Add functionality of copy to clipboard
