+++ WonderfulPalette.xcodeproj/xcuserdata/carter.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -7,7 +7,7 @@
 		<key>WonderfulPalette (iOS).xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>1</integer>
+			<integer>0</integer>
 		</dict>
 		<key>WonderfulPalette (macOS).xcscheme_^#shared#^_</key>
 		<dict>
